Refactor: move specialtx-payload setters/getters to primitives
thus breaking the circular dependency deterministicmns <-> specialtx
