+++ src/evo/specialtx.h
@@ -32,36 +32,6 @@ bool CheckSpecialTxNoContext(const CTransaction& tx, CValidationState& state);
 bool ProcessSpecialTxsInBlock(const CBlock& block, const CBlockIndex* pindex, const CCoinsViewCache* view, CValidationState& state, bool fJustCheck);
 bool UndoSpecialTxsInBlock(const CBlock& block, const CBlockIndex* pindex);
 
-template <typename T>
-inline bool GetTxPayload(const std::vector<unsigned char>& payload, T& obj)
-{
-    CDataStream ds(payload, SER_NETWORK, PROTOCOL_VERSION);
-    try {
-        ds >> obj;
-    } catch (std::exception& e) {
-        return false;
-    }
-    return ds.empty();
-}
-template <typename T>
-inline bool GetTxPayload(const CMutableTransaction& tx, T& obj)
-{
-    return tx.hasExtraPayload() && GetTxPayload(*tx.extraPayload, obj);
-}
-template <typename T>
-inline bool GetTxPayload(const CTransaction& tx, T& obj)
-{
-    return tx.hasExtraPayload() && GetTxPayload(*tx.extraPayload, obj);
-}
-
-template <typename T>
-void SetTxPayload(CMutableTransaction& tx, const T& payload)
-{
-    CDataStream ds(SER_NETWORK, PROTOCOL_VERSION);
-    ds << payload;
-    tx.extraPayload.emplace(ds.begin(), ds.end());
-}
-
 uint256 CalcTxInputsHash(const CTransaction& tx);
 
 #endif // PIVX_SPECIALTX_H
