+++ src/primitives/transaction.h
@@ -452,4 +452,35 @@ template <typename Tx> static inline CTransactionRef MakeTransactionRef(Tx&& txI
 static inline CTransactionRef MakeTransactionRef(const CTransactionRef& txIn) { return txIn; }
 static inline CTransactionRef MakeTransactionRef(CTransactionRef&& txIn) { return std::move(txIn); }
 
+/* Special tx payload handling */
+template <typename T>
+inline bool GetTxPayload(const std::vector<unsigned char>& payload, T& obj)
+{
+    CDataStream ds(payload, SER_NETWORK, PROTOCOL_VERSION);
+    try {
+        ds >> obj;
+    } catch (std::exception& e) {
+        return false;
+    }
+    return ds.empty();
+}
+template <typename T>
+inline bool GetTxPayload(const CMutableTransaction& tx, T& obj)
+{
+    return tx.hasExtraPayload() && GetTxPayload(*tx.extraPayload, obj);
+}
+template <typename T>
+inline bool GetTxPayload(const CTransaction& tx, T& obj)
+{
+    return tx.hasExtraPayload() && GetTxPayload(*tx.extraPayload, obj);
+}
+
+template <typename T>
+void SetTxPayload(CMutableTransaction& tx, const T& payload)
+{
+    CDataStream ds(SER_NETWORK, PROTOCOL_VERSION);
+    ds << payload;
+    tx.extraPayload.emplace(ds.begin(), ds.end());
+}
+
 #endif // BITCOIN_PRIMITIVES_TRANSACTION_H
