+++ test/lint/lint-circular-dependencies.sh
@@ -20,7 +20,6 @@ EXPECTED_CIRCULAR_DEPENDENCIES=(
     "consensus/params -> consensus/upgrades -> consensus/params"
     "crypter -> wallet/wallet -> crypter"
     "evo/deterministicmns -> evo/providertx -> evo/deterministicmns"
-    "evo/deterministicmns -> evo/specialtx -> evo/deterministicmns"
     "evo/deterministicmns -> masternodeman -> evo/deterministicmns"
     "evo/providertx -> evo/specialtx -> evo/providertx"
     "kernel -> validation -> kernel"
@@ -57,6 +56,7 @@ EXPECTED_CIRCULAR_DEPENDENCIES=(
     "chainparamsbase -> util/system -> logging -> chainparamsbase"
     "coins -> policy/fees -> txmempool -> coins"
     "consensus/zerocoin_verify -> spork -> validation -> consensus/zerocoin_verify"
+    "evo/deterministicmns -> evo/providertx -> evo/specialtx -> evo/deterministicmns"
     "evo/deterministicmns -> masternode -> masternode-sync -> evo/deterministicmns"
     "evo/deterministicmns -> masternodeman -> net_processing -> evo/deterministicmns"
     "evo/deterministicmns -> masternode -> wallet/wallet -> evo/deterministicmns"
